chg:dev:#17 Add DDL statements CREATE, DROP, ALTER for DataSource
Added 3 DDL Statements for Datasource:
CREATE DATASOURCE: Creates DATASOURCE
ALTER DATASOURCE: Updates DATASOURCE
DROP DATASOURCE: Drops DATASOURCE

+++ catalog/src/main/java/com/qubole/quark/catalog/db/pojo/DataSource.java
@@ -89,7 +89,7 @@ public class DataSource {
     return dsSetId;
   }
 
-  public void setDsSetId(int dsSetId) {
+  public void setDsSetId(long dsSetId) {
     this.dsSetId = dsSetId;
   }
